refactor: useGetUserList hook 구현
